refactor(pch): replace stdafx.* with pch.* and update project references
Renames StdAfx.h/cpp to pch.h/cpp and updates all includes, vcxproj
PCH settings, and filters to use the modern pch naming convention.
This removes legacy VS6-era stdafx usage and aligns the project with
current MSVC expectations.
