+++ ColorCop.cpp
@@ -4,7 +4,7 @@
 // ColorCop.cpp : Defines the class behaviors for the application.
 //
 
-#include "stdafx.h"
+#include "pch.h"
 #include "ColorCop.h"
 #include "ColorCopDlg.h"
 #include "Defaults.h"
