+++ ColorCop.h
@@ -7,7 +7,7 @@
 #pragma once
 
 #ifndef __AFXWIN_H__
-    #error include 'stdafx.h' before including this file for PCH
+    #error include 'pch.h' before including this file for PCH
 #endif
