+++ pch.cpp
@@ -7,4 +7,4 @@
 // work here, all other source files can compile faster and avoid repeatedly
 // parsing the same large headers.
 
-#include "stdafx.h"
+#include "pch.h"
